ClientCodeGenerationDispatcher_Finds_Solitary_Custom now adds OpenRiaServices.DomainServices.Tools.TextTemplate to list of assemblies to search for code generators
This reproduces a bug in open ria services where, given a custom code generator, the FindCodeGenerator method (incorrectly) finds two (or more) code generators from which it cannot decide which one to use. and therefore returns to the default code generator.
This is different behavior than the latest wcf ria services release.