Bug fix: IsSystemAssembly did not identify Open Ria Services assemblies as system assemblies. In case of custom code generators this resulted in the system not picking up the custom code generator at all.
This fix is to have IsSystemAssembly search for the assembly name in a static list of Open Ria Services assembly names. If the assembly name is found in the array, the assembly is identified as a system assembly.

The problem with this approach is that whenever Open Ria Services is extended with an additional assembly, this array needs to be updated.