Remove "Async" from Task service methods +Fix Tests
Remove ending "Async" from the name of Task-Async service methods (final
part of workitem 41).
I.e. if the method is called EchoAsync then the client will get the two
methods "Echo" and the task returning "EchoAsync", instead of
"EchoAsync" and "EchoAsyncAsync".

Update test for Task-Async methods to reflect the changes and fix some,
minor bugs in the tests.