Added endpoint route pattern functionality to domain services
Implemented a new feature to configure the pattern used for generating endpoint routes in domain services

The `AddDomainService(Type type)` method was modified to use the new `GetDomainServiceRoute(type)` method.

A new file `DomainServiceEndpointRoutePatternAttribute.cs` was added, defining the `EndpointRoutePattern` enum and the `DomainServiceEndpointRoutePatternAttribute` attribute. The `GenerateConstructors()` method in `CSharpDomainContextGenerator.cs` and `CSharpDomainContextGenerator.tt` was modified to use the new `GetDomainServiceUri()` method.

A new method `GetDomainServiceUri()` was added to `CSharpDomainContextGenerator.partial.cs` and `DomainServiceProxyGenerator.cs`. The `ClientCodeGenerationDispatcher.cs` and `ClientCodeGenerationOptions.cs` were modified to load and set the `DefaultEndpointRoutePattern` option. The `DomainServiceProxyGenerator.cs` was modified to use the new `GetDomainServiceUri()` method.

A new file `EndpointRoutePattern.cs` was added, defining the `EndpointRoutePattern` enum. The `Program.cs` was modified to set the `DomainServiceEndpointRoutePattern` attribute for the assembly to `EndpointRoutePattern.WCF`.