Updated BinaryHttpDomainClientFactory and Main method
Updated the `BinaryHttpDomainClientFactory` class in the `OpenRiaServices.Client.DomainClients` namespace. The `_httpClientFactory` field now accepts a `Uri` parameter for `HttpClient` creation.

Updated `DomainClientFactory` so that it will only append "/binary/" for WCF style Uris, for all other kinds of uri it will ensure there is a trailing "/" instead.

Add special workaround for AspNetCore tests so that they use the "Fullname" endpoint pattern even though the client is generated to use WCF style pattern.